Refactor database settings pages and init SSl configuration UI
