+++ studio/components/interfaces/Settings/Database/NetworkRestrictions/NetworkRestrictions.tsx
@@ -40,7 +40,7 @@ const NetworkRestrictions = ({}) => {
             title="Network Restrictions"
             description="Allow specific IP ranges to have access to your database."
           />
-          <div className="flex items-center space-x-2">
+          <div className="flex items-center space-x-2 mb-6">
             <Link href="https://supabase.com/docs/guides/platform/network-restrictions">
               <a target="_blank">
                 <Button type="default" icon={<IconExternalLink />}>
