+++ studio/components/ui/Forms/FormSection.tsx
@@ -39,7 +39,7 @@ const FormSectionLabel = ({
 }) => {
   if (description !== undefined) {
     return (
-      <div className={`flex flex-col space-y-4 col-span-12 lg:col-span-5 ${className}`}>
+      <div className={`flex flex-col space-y-2 col-span-12 lg:col-span-5 ${className}`}>
         <label className="text-scale-1200 text-sm">{children}</label>
         {description}
       </div>
